% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.R
\name{id}
\alias{id}
\title{Generic Function for Extracting IDs}
\usage{
id(object, ...)
}
\arguments{
\item{object}{An object from which to extract the ID attribute.}

\item{...}{Additional arguments.}
}
\value{
A vector of IDs representing the cluster allocations for the data.
}
\description{
A generic function to extract the ID attribute from an object.
}
